refactor(internal/librarianops): simplify generate command interface
Remove the --all flag and simplify the -C flag behavior. The repo name
is now inferred from filepath.Base(dir) when using -C, eliminating the
need to specify it separately.

The new interface is:

  librarianops generate <repo>
  librarianops generate -C <dir>

The --all flag is removed, since the expectation is that we would only
ever process on repo at at time locally, or can run:

   librarian generate google-cloud-rust && librarian generate google-cloud-go && ...

Also move generate-specific code to generate.go and generate_test.go.
